# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AP Automation API",
    "description": "API for accounts payable automation with ERP-neutral extraction and configurable connectors",
    "version": "1.0.0"
  },
  "paths": {
...
    "/connectors/{connector_type}": {
      "get": {
        "tags": [
          "Connectors"
        ],
        "summary": "Get Connector Info",
        "description": "Get information about a specific connector.",
        "operationId": "get_connector_info_connectors__connector_type__get",
        "parameters": [
          {
            "name": "connector_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connector Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/connectors/{connector_type}/configure": {
      "post": {
        "tags": [
          "Connectors"
        ],
        "summary": "Configure Connector",
        "description": "Configure an ERP connector.\n\nFor Business Central, auth_config should include:\n- tenant_id: Azure AD tenant ID\n- client_id: Application client ID\n- client_secret: Client secret\n\ncustom_settings can include:\n- environment: BC environment name (default: \"production\")\n- api_version: API version (default: \"v2.0\")",
        "operationId": "configure_connector_connectors__connector_type__configure_post",
        "parameters": [
          {
            "name": "connector_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connector Type"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectorConfigRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Connectors"
        ],
        "summary": "Remove Configuration",
        "description": "Remove connector configuration.",
        "operationId": "remove_configuration_connectors__connector_type__configure_delete",
        "parameters": [
          {
            "name": "connector_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connector Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Remove Configuration Connectors  Connector Type  Configure Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/connectors/{connector_type}/test": {
      "post": {
        "tags": [
          "Connectors"
        ],
        "summary": "Test Connection",
        "description": "Test the connection to the ERP system.",
        "operationId": "test_connection_connectors__connector_type__test_post",
        "parameters": [
          {
            "name": "connector_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connector Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionTestResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/connectors/{connector_type}/entities/{entity_type}": {
      "get": {
        "tags": [
          "Connectors"
        ],
        "summary": "List Erp Entities",
        "description": "List entities (vendors, GL accounts, etc.) from the ERP.\n\nEntity types: vendor, gl_account, location, dimension",
        "operationId": "list_erp_entities_connectors__connector_type__entities__entity_type__get",
        "parameters": [
          {
            "name": "connector_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connector Type"
            }
          },
          {
            "name": "entity_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Type"
            }
          },
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Active Only"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Erp Entities Connectors  Connector Type  Entities  Entity Type  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/connectors/{connector_type}/lookup/{entity_type}": {
      "get": {
        "tags": [
          "Connectors"
        ],
        "summary": "Lookup Entity",
        "description": "Lookup a specific entity in the ERP.",
        "operationId": "lookup_entity_connectors__connector_type__lookup__entity_type__get",
        "parameters": [
          {
            "name": "connector_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connector Type"
            }
          },
          {
            "name": "entity_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Type"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Lookup Entity Connectors  Connector Type  Lookup  Entity Type  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectionTestResult": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "ConnectionTestResult",
        "description": "Result of a connection test."
      },
      "ConnectorConfigRequest": {
        "properties": {
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "company_id": {
            "type": "string",
            "title": "Company Id",
            "description": "Company/tenant identifier in ERP"
          },
          "auth_config": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Auth Config",
            "description": "Authentication configuration (varies by connector type)"
          },
          "custom_settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Custom Settings",
            "description": "Connector-specific settings"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "auth_config"
        ],
        "title": "ConnectorConfigRequest",
        "description": "Request to configure a connector."
      },
      "ConnectorConfigResponse": {
        "properties": {
          "connector_type": {
            "type": "string",
            "title": "Connector Type"
          },
          "status": {
            "$ref": "#/components/schemas/ConnectorStatus"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "connector_type",
          "status",
          "message"
        ],
        "title": "ConnectorConfigResponse",
        "description": "Response after configuring a connector."
      },
      "ConnectorInfo": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/ConnectorStatus"
          },
          "is_configured": {
            "type": "boolean",
            "title": "Is Configured"
          },
          "last_connected": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Connected"
          }
        },
        "type": "object",
        "required": [
          "type",
          "name",
          "description",
          "status",
          "is_configured"
        ],
        "title": "ConnectorInfo",
        "description": "Information about an available connector."
      },
      "ConnectorStatus": {
        "type": "string",
        "enum": [
          "configured",
          "not_configured",
          "connected",
          "disconnected",
          "error"
        ],
        "title": "ConnectorStatus",
        "description": "Connector status."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}